python/requirements.txt
Django==1.8
argparse==1.2.1
djangorestframework==3.4.7
wget==3.2
wsgiref==0.1.2
